pyproject.toml
[project]
name = "finance-backend"
version = "1.0.0"
description = "Production-quality FastAPI backend for a finance dashboard."
readme = "README.md"
requires-python = ">=3.11,<4.0"
authors = [{ name = "OpenAI Codex", email = "codex@openai.com" }]
dependencies = [
    "fastapi==0.115.12",
    "uvicorn[standard]==0.34.0",
    "sqlalchemy==2.0.38",
    "alembic==1.15.1",
    "pydantic[email]==2.10.6",
    "pydantic-settings==2.7.1",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.0.1",
    "python-multipart==0.0.20",
    "python-dotenv==1.0.1",
]

[tool.poetry]
packages = [{ include = "app" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
